+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Copy and pasting the git commit messages is __NOT__ enough.
 ### Changed
 - `history_vehicles_replacement_for_imitation_learning.py` now uses new Imitation action space.  See Issue #844.
 - Vehicles with a `BoxChassis` can now use an `AccelerometerSensor` too.
+- When importing NGSIM history data, vehicle speeds are recomputed.
 ### Fixed
 - Allow for non-dynamic action spaces to have action controllers.  See PR #854.
 - Fixed a minor bug in `sensors.py` which triggered `wrong_way` event when the vehicle goes into an intersection. See Issue #846.